maf_reasoning_effort/requirements.txt
agent-framework>=1.0.0
azure-identity>=1.21.0
python-dotenv>=1.0.1
rich>=13.7.0
